requirements.txt

PyQt6
openai
python-dotenv
